requirements-stdborg.txt
aiofiles
aiohttp
beautifulsoup4
cfscrape
coffeehouse
covid
cowpy
cryptg
csvfaker
colour
dadjokes-cli
deezloader
dnspython
emoji
fast.com
geopy
gitpython
glitch_this
google-api-python-client
google_auth_oauthlib
googletrans
gsearch>=1.6.0
gtts
git+https://github.com/bisoncorps/search-engine-parser
hachoir
heroku3
html_telegraph_poster
htmlement
httplib2==0.15.0
humanize
hurry.filesize
jikanpy
jotquote
justwatch
lottie
lxml
lyricsgenius
names
nekos.py
oauth2client
Pillow
PyGithub
pathlib
patool
pendulum
psutil
psycopg2-binary
pySmartDL
pybase64
pyfiglet
pymongo
python-barcode
python-magic
pytube3
pygments
pytuneteller
qrcode
random_cli
regex
remotezip
requests
selenium
spamwatch
speedtest-cli
spotify_token
sqlalchemy
telegraph
urbandict
validators
wget
wheel
wikipedia
youtube-dl